refactor: move AppLifetimeService to AppLifetime namespace
